Change stack to fix program counter overflow bug

+++ include/tinybasic/stack.h
@@ -5,9 +5,9 @@
 enum { STACK_SIZE = 16, };\
 
 struct stack {
-    uint8_t data[STACK_SIZE];
+    int data[STACK_SIZE];
     int pos;
 };
 
-int stack_push(struct stack *stack, uint8_t value);
-int stack_pop(struct stack *stack, uint8_t *value);
+int stack_push(struct stack *stack, int value);
+int stack_pop(struct stack *stack, int *value);
